Extracted the prepare command and removed the check for latest
